+++ src/main/java/io/github/zivasd/spring/boot/data/shared/query/DeciderParam.java
@@ -12,5 +12,5 @@ import java.lang.annotation.Target;
 public @interface DeciderParam {
     String value() default "";
 
-    boolean bindable() default false;
+    boolean bindable() default true;
 }